Most data is now validated on input. It's validated once in the resource classes and once in the EntityDtoMapper.
The exception would be when creating an entity and the parentId is required in the request body. Validating that parentId would require a more complex validation mechanism. When you create a child entity and pass a negative parentId in the request body, it will still throw a 500 internal server error that is handled by the generic exception handler, but it will not respond with a detailed list of constriant violations.