Minimum Meeting Time can now be adjusted by the client.
The minimum meeting must now be specified by the client when attempting to find event times using the query parameter minimum. The value is specified in minutes. It cannot be null, negative or greater than 960. The client should specify a default value of 15 if no other value is selected. Overlaps of 10 minutes or less will not be considered when finding event times.